lots of stuff
Rearranged environment again. Moved the variables into an encapsulated
class so the values share amoung all classes the module is included in.
Same with assertion statistics.
Added tests for base, moved tdd_deploy tests to a sane place, added
color to test results, renamed a lot of stuff.
It is a mess - but getting closer